@@ -1,20 +1,19 @@
 {
   "name": "hofff/contao-content-navigation",
   "description": "Provides a table of contents for elements based on one or more articles within the page.",
+  "license": "LGPL-3.0-or-later",
+  "type": "contao-bundle",
   "keywords": [
     "contao",
     "contents",
     "nabigation",
     "toc"
   ],
-  "type": "contao-bundle",
-  "license": "LGPL-3.0-or-later",
-  "homepage": "https://www.hofff.com",
   "authors": [
     {
       "name": "Nicky Hoff",
-      "homepage": "https://www.hofff.com",
       "email": "nick@hofff.com",
+      "homepage": "https://www.hofff.com",
       "role": "Manager"
     },
     {
@@ -24,29 +23,43 @@
       "role": "Developer"
     }
   ],
+  "homepage": "https://www.hofff.com",
   "require": {
-    "php": "^7.1",
+    "php": "^7.2 || ^8.0",
     "ext-pdo": "*",
-    "contao/core-bundle": "^4.4",
+    "ausi/slug-generator": "^1.1",
+    "contao/core-bundle": "^4.9",
+    "doctrine/dbal": "^2.11 || ^3.0",
     "patchwork/utf8": "^1.2",
-    "symfony/http-kernel": "^3.3 || ^4.0",
-    "symfony/config": "^3.3 || ^4.0",
-    "symfony/dependency-injection": "^3.3 || ^4.0",
-    "doctrine/dbal": "^2.5",
-    "ausi/slug-generator": "^1.1"
+    "symfony/config": "^4.4 || ^5.1",
+    "symfony/dependency-injection": "^4.4 || ^5.1",
+    "symfony/http-kernel": "^4.4 || ^5.1"
   },
   "require-dev": {
-    "contao/manager-plugin": "^2.0"
+    "contao/manager-plugin": "^2.1",
+    "doctrine/coding-standard": "^9.0.",
+    "phpcq/runner-bootstrap": "^1.0@dev"
+  },
+  "conflict": {
+    "contao/manager-plugin": "<2.1 || >=3.0"
   },
   "autoload": {
     "psr-4": {
       "Hofff\\Contao\\ContentNavigation\\": "src/"
     }
   },
+  "config": {
+    "allow-plugins": {
+      "contao-components/installer": false,
+      "contao/manager-plugin": true,
+      "dealerdirect/phpcodesniffer-composer-installer": true
+    },
+    "sort-packages": true
+  },
   "extra": {
-    "contao-manager-plugin": "Hofff\\Contao\\ContentNavigation\\ContaoManager\\Plugin",
     "branch-alias": {
       "dev-master": "1.1.x-dev"
-    }
+    },
+    "contao-manager-plugin": "Hofff\\Contao\\ContentNavigation\\ContaoManager\\Plugin"
   }
 }
