@@ -4,9 +4,6 @@ declare(strict_types=1);
 
 /**
  * Contao Content Navigation
- *
- * @copyright 2010-2011 InfinitySoft
- * @copyright 2018 nickname. Büro für visuelle Kommunikation Nicky Hoff
  */
 
 $GLOBALS['TL_LANG']['tl_content']['toc_legend'] = 'Table of contents';
@@ -14,13 +11,13 @@ $GLOBALS['TL_LANG']['tl_content']['toc_legend'] = 'Table of contents';
 /*
  * Fields
  */
-$GLOBALS['TL_LANG']['tl_content']['hofff_toc_source']            = array('Section', 'Select the section for which you want to create the content navigation.');
-$GLOBALS['TL_LANG']['tl_content']['hofff_toc_min_level']         = array('Start level', 'Start at this headline level (inclusive).');
-$GLOBALS['TL_LANG']['tl_content']['hofff_toc_max_level']         = array('Stop level', 'Stop at this headline level (inclusive).');
-$GLOBALS['TL_LANG']['tl_content']['hofff_toc_include']           = array('Include in content navigation', 'Element will be included in the content navigation if a headline is defined. If there is no CSS ID, it will be generated automatically.');
-$GLOBALS['TL_LANG']['tl_content']['hofff_toc_force_request_uri'] = array('Force request uri', 'Use current request uri as base link instead of related page');
-$GLOBALS['TL_LANG']['tl_content']['hofff_toc_source_column'] = 'Article in column';
-$GLOBALS['TL_LANG']['tl_content']['hofff_toc_source_page']   = 'Article in page';
+$GLOBALS['TL_LANG']['tl_content']['hofff_toc_source']            = ['Section', 'Select the section for which you want to create the content navigation.'];
+$GLOBALS['TL_LANG']['tl_content']['hofff_toc_min_level']         = ['Start level', 'Start at this headline level (inclusive).'];
+$GLOBALS['TL_LANG']['tl_content']['hofff_toc_max_level']         = ['Stop level', 'Stop at this headline level (inclusive).'];
+$GLOBALS['TL_LANG']['tl_content']['hofff_toc_include']           = ['Include in content navigation', 'Element will be included in the content navigation if a headline is defined. If there is no CSS ID, it will be generated automatically.'];
+$GLOBALS['TL_LANG']['tl_content']['hofff_toc_force_request_uri'] = ['Force request uri', 'Use current request uri as base link instead of related page'];
+$GLOBALS['TL_LANG']['tl_content']['hofff_toc_source_column']     = 'Article in column';
+$GLOBALS['TL_LANG']['tl_content']['hofff_toc_source_page']       = 'Article in page';
 
 /*
  * Other
